Move QueryType to query-params
